requirements.txt
hydra-core
omegaconf
torch
torchvision
